+++ apps/demo/src/demo-data/pbApplicationsData.ts
@@ -1,0 +1,10 @@
+import { PBApplications } from '@digital-ai/dot-components';
+
+export const pbApplications: PBApplications = {
+  '5ba79d5929799324b10f1111': {
+    waitingMessage: 'Waiting for commits',
+  },
+  '5ba79d5929799324b10f1222': {
+    waitingMessage: 'Commits are not tracked',
+  },
+};