+++ libs/dot-components/src/lib/ctm/progression-board/application/data/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_APP_WAITING_MESSAGE =
+  'To see stories and defects here, configure a source control webhook, make a commit and use the Track code changes task to update the board.';