+++ libs/dot-components/src/lib/ctm/progression-board/phase/WaitingPhase.tsx
@@ -3,11 +3,17 @@ import { CommonProps } from '../../../components/CommonProps';
 import { useStylesWithRootClass } from '../../../components/useStylesWithRootClass';
 import { DotIcon, DotTypography } from '../../../components';
 import { rootClassName, StyledWaitingPhase } from './WaitingPhase.styles';
+import { DEFAULT_APP_WAITING_MESSAGE } from '../application/data/constants';
+
+export interface WaitingPhaseProps extends CommonProps {
+  waitingMessage: string;
+}
 
 export const WaitingPhase = ({
   className,
   'data-testid': dataTestId,
-}: CommonProps) => {
+  waitingMessage = DEFAULT_APP_WAITING_MESSAGE,
+}: WaitingPhaseProps) => {
   const rootClasses = useStylesWithRootClass(rootClassName, className);
 
   return (
@@ -18,10 +24,7 @@ export const WaitingPhase = ({
           data-testid={`${dataTestId}-waiting-icon`}
           iconId="pending-clock"
         />
-        <DotTypography variant="body2">
-          To see stories and defects here, configure a source control webhook,
-          make a commit and use the Track code changes task to update the board.
-        </DotTypography>
+        <DotTypography variant="body2">{waitingMessage}</DotTypography>
       </div>
     </StyledWaitingPhase>
   );